requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.7
WTForms==3.0.1
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.3
Flask-WTF==1.1.1
fonttools==4.42.1
Genderize==0.3.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.1
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.3
Flask-WTF==1.1.1
fonttools==4.42.1
Genderize==0.3.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.1
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
zope.interface==6.0
